+++ spec/v1/providers/tasks.spec.ts
@@ -28,6 +28,7 @@ import { MockRequest } from "../../fixtures/mockrequest";
 import { runHandler } from "../../helper";
 import { MINIMAL_V1_ENDPOINT } from "../../fixtures";
 import { MINIMIAL_TASK_QUEUE_TRIGGER } from "./fixtures";
+import { runWith } from "../../../src/v1";
 
 describe("#onDispatch", () => {
   it("should return a trigger/endpoint with appropriate values", () => {
@@ -67,6 +68,7 @@ describe("#onDispatch", () => {
       ...MINIMAL_V1_ENDPOINT,
       platform: "gcfv1",
       taskQueueTrigger: {
+        ...MINIMIAL_TASK_QUEUE_TRIGGER,
         rateLimits: {
           maxConcurrentDispatches: 30,
           maxDispatchesPerSecond: 40,
@@ -77,6 +79,35 @@ describe("#onDispatch", () => {
           maxBackoffSeconds: 20,
           maxDoublings: 3,
           minBackoffSeconds: 5,
+        },
+        invoker: ["private"],
+      },
+    });
+  });
+
+  it("should return an endpoint with appropriate values with preserveExternalChanges set", () => {
+    const result = runWith({ preserveExternalChanges: true })
+      .tasks.taskQueue({
+        rateLimits: {
+          maxConcurrentDispatches: 30,
+        },
+        retryConfig: {
+          maxAttempts: 5,
+          maxRetrySeconds: 10,
+        },
+        invoker: "private",
+      })
+      .onDispatch(() => undefined);
+
+    expect(result.__endpoint).to.deep.equal({
+      platform: "gcfv1",
+      taskQueueTrigger: {
+        rateLimits: {
+          maxConcurrentDispatches: 30,
+        },
+        retryConfig: {
+          maxAttempts: 5,
+          maxRetrySeconds: 10,
         },
         invoker: ["private"],
       },
@@ -114,6 +145,10 @@ describe("#onDispatch", () => {
         ...MINIMIAL_TASK_QUEUE_TRIGGER,
         retryConfig: {
           maxAttempts: 5,
+          maxBackoffSeconds: functions.RESET_VALUE,
+          maxDoublings: functions.RESET_VALUE,
+          maxRetrySeconds: functions.RESET_VALUE,
+          minBackoffSeconds: functions.RESET_VALUE,
         },
       },
     });