+++ src/v2/providers/eventarc.ts
@@ -63,9 +63,14 @@ export interface EventarcTriggerOptions extends options.EventHandlerOptions {
   filters?: Record<string, string>;
 
   /**
+   * If true, do not deploy or emulate this function.
+   */
+  omit?: boolean | Expression<boolean>;
+
+  /**
    * Region where functions should be deployed.
    */
-  region?: options.SupportedRegion | string;
+  region?: options.SupportedRegion | string | Expression<string> | ResetValue;
 
   /**
    * Amount of memory to allocate to a function.
@@ -73,7 +78,7 @@ export interface EventarcTriggerOptions extends options.EventHandlerOptions {
   memory?: options.MemoryOption | Expression<number> | ResetValue;
 
   /**
-   * Timeout for the function in sections, possible values are 0 to 540.
+   * Timeout for the function in seconds, possible values are 0 to 540.
    * HTTPS functions can specify a higher timeout.
    *
    * @remarks
@@ -125,7 +130,7 @@ export interface EventarcTriggerOptions extends options.EventHandlerOptions {
   /**
    * Connect cloud function to specified VPC connector.
    */
-  vpcConnector?: string | ResetValue;
+  vpcConnector?: string | Expression<string> | ResetValue;
 
   /**
    * Egress settings for VPC connector.
@@ -135,7 +140,7 @@ export interface EventarcTriggerOptions extends options.EventHandlerOptions {
   /**
    * Specific service account for the function to run as.
    */
-  serviceAccount?: string | ResetValue;
+  serviceAccount?: string | Expression<string> | ResetValue;
 
   /**
    * Ingress settings which control where this function can be called from.
@@ -219,6 +224,13 @@ export function onCustomEventPublished<T = any>(
   copyIfPresent(endpoint.eventTrigger, opts, "retry");
 
   func.__endpoint = endpoint;
+
+  func.__requiredAPIs = [
+    {
+      api: "eventarcpublishing.googleapis.com",
+      reason: "Needed for custom event functions",
+    },
+  ];
 
   return func;
 }